Revert "backend: split DataManager"
This reverts commit 770d8816865dd74d7e8da955a305c840447bd2b8.
