+++ dataland-backend/src/main/kotlin/org/dataland/datalandbackend/controller/MetaDataController.kt
@@ -1,7 +1,7 @@
 package org.dataland.datalandbackend.controller
 
 import org.dataland.datalandbackend.api.MetaDataApi
-import org.dataland.datalandbackend.interfaces.DataProcessorInterface
+import org.dataland.datalandbackend.interfaces.DataManagerInterface
 import org.dataland.datalandbackend.model.DataMetaInformation
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.http.ResponseEntity
@@ -15,7 +15,7 @@ import org.springframework.web.bind.annotation.RestController
 
 @RestController
 class MetaDataController(
-    @Autowired var dataManager: DataProcessorInterface,
+    @Autowired var dataManager: DataManagerInterface,
 ) : MetaDataApi {
 
     override fun getListOfDataMetaInfo(companyId: String?, dataType: String?):
