+++ dataland-backend/src/main/kotlin/org/dataland/datalandbackend/interfaces/DataManagerInterface.kt
@@ -1,0 +1,6 @@
+package org.dataland.datalandbackend.interfaces
+
+/**
+ * Defines the required functionalities for the Dataland data manager
+ */
+interface DataManagerInterface : CompanyDataStoreInterface, DataProcessorInterface, DataStoreInterface
