+++ dataland-backend/src/main/kotlin/org/dataland/datalandbackend/interfaces/DataProcessorInterface.kt
@@ -1,8 +1,6 @@
 package org.dataland.datalandbackend.interfaces
 
-import org.dataland.datalandbackend.model.DataManagerInputToGetData
 import org.dataland.datalandbackend.model.DataMetaInformation
-import org.dataland.datalandbackend.model.StorableDataSet
 
 /**
  * Defines the required functionalities to process data for data manager
@@ -21,17 +19,4 @@ interface DataProcessorInterface {
         companyId: String = "",
         dataType: String = ""
     ): List<DataMetaInformation>
-    /**
-     * Method to make the data manager add data to a data store and store meta data in Dataland
-     * @param storableDataSet contains all the inputs needed by Dataland
-     * @return ID of the newly stored data in the data store
-     */
-    fun addDataSet(storableDataSet: StorableDataSet): String
-
-    /**
-     * Method to make the data manager get the data of a single entry from the data store
-     * @param dataManagerInputToGetData contains all the inputs needed by Dataland
-     * @return data associated with the data ID provided in the input
-     */
-    fun getData(dataManagerInputToGetData: DataManagerInputToGetData): String
 }
