+++ dataland-backend/src/main/kotlin/org/dataland/datalandbackend/service/DataManager.kt
@@ -1,38 +1,51 @@
 package org.dataland.datalandbackend.service
 
+import org.dataland.datalandbackend.edcClient.api.DefaultApi
+import org.dataland.datalandbackend.interfaces.DataManagerInterface
 import org.dataland.datalandbackend.model.CompanyMetaInformation
+import org.dataland.datalandbackend.model.DataManagerInputToGetData
 import org.dataland.datalandbackend.model.DataMetaInformation
+import org.dataland.datalandbackend.model.StorableDataSet
+import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Component
 
 /**
- * Implementation of a data manager for Dataland including metadata storages
+ * Implementation of a data manager for Dataland including meta data storages
  */
 @Component("DefaultManager")
-class DataManager {
+class DataManager(
+    @Autowired var edcClient: DefaultApi
+) : DataManagerInterface {
     var dataMetaData = mutableMapOf<String, DataMetaInformation>()
     var companyData = mutableMapOf<String, CompanyMetaInformation>()
-    internal var companyCounter = 0
+    private var companyCounter = 0
 
-    internal fun verifyCompanyIdExists(companyId: String) {
+    /*
+    ________________________________
+    Helper-methods:
+    ________________________________
+     */
+
+    private fun verifyCompanyIdExists(companyId: String) {
         if (!companyData.containsKey(companyId)) {
             throw IllegalArgumentException("Dataland does not know the company ID $companyId.")
         }
     }
 
-    internal fun verifyDataIdExists(dataId: String) {
+    private fun verifyDataIdExists(dataId: String) {
         if (!dataMetaData.containsKey(dataId)) {
             throw IllegalArgumentException("Dataland does not know the data ID: $dataId.")
         }
     }
 
-    internal fun verifyDataTypeExists(dataType: String) {
+    private fun verifyDataTypeExists(dataType: String) {
         val matchesForDataType = dataMetaData.any { it.value.dataType == dataType }
         if (!matchesForDataType) {
             throw IllegalArgumentException("Dataland does not know the data type: $dataType")
         }
     }
 
-    internal fun verifyDataIdExistsAndIsOfType(dataId: String, dataType: String) {
+    private fun verifyDataIdExistsAndIsOfType(dataId: String, dataType: String) {
         verifyDataIdExists(dataId)
         if (dataMetaData[dataId]!!.dataType != dataType) {
             throw IllegalArgumentException(
@@ -41,5 +54,102 @@ class DataManager {
                     " type $dataType."
             )
         }
+    }
+
+    /*
+    ________________________________
+    Methods to route data inserts and queries to the data store and save meta data in the Dataland-Meta-Data-Storage:
+    ________________________________
+     */
+
+    override fun addDataSet(storableDataSet: StorableDataSet): String {
+        verifyCompanyIdExists(storableDataSet.companyId)
+
+        val dataId = edcClient.insertData(storableDataSet.data)
+
+        if (dataMetaData.containsKey(dataId)) {
+            throw IllegalArgumentException("The data ID $dataId already exists in Dataland.")
+        }
+
+        val dataMetaInformation =
+            DataMetaInformation(dataId, dataType = storableDataSet.dataType, companyId = storableDataSet.companyId)
+        dataMetaData[dataId] = dataMetaInformation
+        companyData[storableDataSet.companyId]!!.dataRegisteredByDataland.add(dataMetaInformation)
+        return dataId
+    }
+
+    override fun getData(dataManagerInputToGetData: DataManagerInputToGetData): String {
+        verifyDataIdExistsAndIsOfType(dataManagerInputToGetData.dataId, dataManagerInputToGetData.dataType)
+
+        val data = edcClient.selectDataById(dataManagerInputToGetData.dataId)
+
+        if (data == "") {
+            throw IllegalArgumentException(
+                "No data set with the id: ${dataManagerInputToGetData.dataId} " +
+                    "could be found in the data store."
+            )
+        }
+        return data
+    }
+
+    /*
+    ________________________________
+    Methods to process meta data:
+    ________________________________
+     */
+
+    override fun searchDataMetaInfo(dataId: String, companyId: String, dataType: String): List<DataMetaInformation> {
+        if (dataId.isNotEmpty()) {
+            verifyDataIdExists(dataId)
+            return listOf(dataMetaData[dataId]!!)
+        }
+
+        var matches: Map<String, DataMetaInformation> = dataMetaData
+
+        if (companyId.isNotEmpty()) {
+            verifyCompanyIdExists(companyId)
+            matches = matches.filter { it.value.companyId == companyId }
+        }
+        if (dataType.isNotEmpty()) {
+            verifyDataTypeExists(dataType)
+            matches = matches.filter { it.value.dataType == dataType }
+        }
+
+        return matches.map {
+            DataMetaInformation(dataId = it.key, dataType = it.value.dataType, companyId = it.value.companyId)
+        }
+    }
+
+    /*
+    ________________________________
+    Methods to add and retrieve company info and save associated company meta data in the
+    Dataland-Company-Meta-Data-Storage:
+    ________________________________
+     */
+
+    override fun addCompany(companyName: String): CompanyMetaInformation {
+        companyCounter++
+        companyData["$companyCounter"] = CompanyMetaInformation(
+            companyId = companyCounter.toString(),
+            companyName = companyName,
+            dataRegisteredByDataland = mutableListOf()
+        )
+        return companyData["$companyCounter"]!!
+    }
+
+    override fun listCompaniesByName(companyName: String): List<CompanyMetaInformation> {
+        return companyData.filter { it.value.companyName.contains(companyName, true) }.map {
+            CompanyMetaInformation(
+                companyId = it.key,
+                companyName = it.value.companyName,
+                dataRegisteredByDataland = it.value.dataRegisteredByDataland
+            )
+        }
+    }
+
+    override fun getCompanyById(companyId: String): CompanyMetaInformation {
+        verifyCompanyIdExists(companyId)
+
+        return companyData[companyId]!!
     }
 }
