+++ dataland-backend/src/test/kotlin/org/dataland/datalandbackend/service/DataManagerTest.kt
@@ -1,19 +1,27 @@
 package org.dataland.datalandbackend.service
 
+import org.dataland.datalandbackend.edcClient.api.DefaultApi
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
 import org.springframework.boot.test.context.SpringBootTest
 
 @SpringBootTest
-class CompanyDataManagerTest {
+class DataManagerTest {
 
-    val testCompanyDataManager = CompanyDataManager()
+    val testManager = DataManager(edcClient = DefaultApi(basePath = "dummy"))
+
     val testCompanyNamesToStore = listOf("Imaginary-Company_I", "Fantasy-Company_II", "Dream-Company_III")
+
+    /*
+    ________________________________
+    Tests for all data manager functionalities associated with adding and getting companies:
+    ________________________________
+     */
 
     @Test
     fun `add the first company and check if its name is as expected by using the return value of addCompany`() {
-        val companyMetaInformation = testCompanyDataManager.addCompany(testCompanyNamesToStore[0])
+        val companyMetaInformation = testManager.addCompany(testCompanyNamesToStore[0])
         assertEquals(
             companyMetaInformation.companyName, testCompanyNamesToStore[0],
             "The company name in the post-response does not match the actual name of the company to be posted."
@@ -23,10 +31,10 @@ class CompanyDataManagerTest {
     @Test
     fun `add all companies then retrieve them as a list and check for each company if its name is as expected`() {
         for (companyName in testCompanyNamesToStore) {
-            testCompanyDataManager.addCompany(companyName)
+            testManager.addCompany(companyName)
         }
 
-        val allCompaniesInStore = testCompanyDataManager.listCompaniesByName("")
+        val allCompaniesInStore = testManager.listCompaniesByName("")
         for ((counter, storedCompany) in allCompaniesInStore.withIndex()) {
             assertEquals(
                 testCompanyNamesToStore[counter], storedCompany.companyName,
@@ -38,11 +46,11 @@ class CompanyDataManagerTest {
     @Test
     fun `add all companies and search for them one by one by using their names`() {
         for (companyName in testCompanyNamesToStore) {
-            testCompanyDataManager.addCompany(companyName)
+            testManager.addCompany(companyName)
         }
 
         for (companyName in testCompanyNamesToStore) {
-            val searchResponse = testCompanyDataManager.listCompaniesByName(companyName)
+            val searchResponse = testManager.listCompaniesByName(companyName)
             assertEquals(
                 companyName, searchResponse.first().companyName,
                 "The posted company could not be found in the data store by searching for its name."
@@ -53,7 +61,14 @@ class CompanyDataManagerTest {
     @Test
     fun `get companies with a name that does not exist`() {
         assertThrows<IllegalArgumentException> {
-            testCompanyDataManager.listCompaniesByName("error")
+            testManager.listCompaniesByName("error")
+        }
+    }
+
+    @Test
+    fun `get the data sets for a company id that does not exist`() {
+        assertThrows<IllegalArgumentException> {
+            testManager.searchDataMetaInfo(companyId = "error")
         }
     }
 }
